Development

+++ GoalMogul-iOS/src/Main/Goal/GoalCard/GoalCard.js
@@ -71,8 +71,20 @@ import StepIcon from '../../../asset/utils/steps.png';
 import ProgressBarLarge from '../../../asset/utils/progressBar_large.png';
 import ProgressBarLargeCounter from '../../../asset/utils/progressBar_counter_large.png';
 
+// Actions
+import { openGoalDetail } from '../../../redux/modules/home/mastermind/actions';
+import {
+  shareGoalToMastermind
+} from '../../../redux/modules/goal/GoalDetailActions';
+
 // Constants
 import { IPHONE_MODELS } from '../../../Utils/Constants';
+
+// Utils
+import { makeCaretOptions } from '../../../redux/middleware/utils';
+
+// Styles
+import { APP_BLUE } from '../../../styles';
 
 const { height } = Dimensions.get('window');
 const CardHeight = height * 0.7;
@@ -180,6 +192,37 @@ class GoalCard extends React.PureComponent {
     });
   };
 
+  // Original style 1 currently used
+  // buttonStyle={{
+  //   selected: {
+  //     backgroundColor: '#f8f8f8',
+  //     tintColor: '#1998c9',
+  //     color: '#1998c9',
+  //     fontWeight: '600'
+  //   },
+  //   unselected: {
+  //     backgroundColor: 'white',
+  //     tintColor: '#696969',
+  //     color: '#696969',
+  //     fontWeight: '600'
+  //   }
+  // }} 
+
+  // Original style 2
+  // buttonStyle={{
+  //   selected: {
+  //     backgroundColor: APP_BLUE,
+  //     tintColor: 'white',
+  //     color: 'white',
+  //     fontWeight: '700'
+  //   },
+  //   unselected: {
+  //     backgroundColor: 'white',
+  //     tintColor: '#616161',
+  //     color: '#616161',
+  //     fontWeight: '600'
+  //   }
+  // }}   
   _renderHeader = props => {
     return (
       <TabButtonGroup 
@@ -198,7 +241,7 @@ class GoalCard extends React.PureComponent {
             color: '#696969',
             fontWeight: '600'
           }
-        }}  
+        }} 
       />
     );
   };
@@ -275,13 +318,45 @@ class GoalCard extends React.PureComponent {
 
     const pageId = _.get(PAGE_TYPE_MAP, 'goalFeed');
 
+    const selfOptions = makeCaretOptions('Goal', item);
+
     const caret = {
       self: {
-        options: [{ option: 'Delete' }],
-        onPress: () => {
-          return this.props.deleteGoal(_id, pageId);
+        options: selfOptions,
+        onPress: (key) => {
+          if (key === 'Delete') {
+            return this.props.deleteGoal(_id, pageId);
+          }
+
+          let initialProps = {};
+          if (key === 'Edit Goal') {
+            initialProps = { initialShowGoalModal: true };
+            this.props.openGoalDetail(item, initialProps);
+            return;
+          }
+          if (key === 'Share to Goal Feed') {
+            // It has no pageId so it won't have loading animation
+            return this.props.shareGoalToMastermind(_id);
+          }
+          if (key === 'Mark as Complete') {
+            initialProps = { 
+              initialMarkGoalAsComplete: true,
+              refreshGoal: false
+            };
+            this.props.openGoalDetail(item, initialProps);
+            return;
+          }
+
+          if (key === 'Unmark as Complete') {
+            initialProps = { 
+              initialUnMarkGoalAsComplete: true,
+              refreshGoal: false
+            };
+            this.props.openGoalDetail(item, initialProps);
+            return;
+          }
         },
-        shouldExtendOptionLength: false
+        shouldExtendOptionLength: owner._id === this.props.userId
       },
       others: {
         options: [
@@ -412,14 +487,15 @@ class GoalCard extends React.PureComponent {
           textStyle={{ color: '#FBDD0D' }}
           iconStyle={{ tintColor: '#FBDD0D', height: 26, width: 26 }}
           onPress={() => {
-            console.log(`${DEBUG_KEY}: user clicks suggest icon`);
+            console.log(`${DEBUG_KEY}: user clicks comment icon`);
             this.props.onPress(
               this.props.item, 
-              // { 
-              //   type: 'Comment', 
-              //   _id: undefined,
-              //   initialShowSuggestionModal: false
-              // }
+              { 
+                type: 'comment', 
+                _id: undefined,
+                initialShowSuggestionModal: false,
+                initialFocusCommentBox: true
+              }
             );
           }}
         />
@@ -452,7 +528,7 @@ class GoalCard extends React.PureComponent {
             </TouchableOpacity>
             { // Disable tabs if neither needs or steps
               _.isEmpty(steps) && _.isEmpty(needs)
-              ? ''
+              ? null
               : (
                 <View style={{ height: tabHeight }}>
                   {this.renderTabs()}
@@ -544,6 +620,8 @@ export default connect(
     chooseShareDest,
     deleteGoal,
     subscribeEntityNotification,
-    unsubscribeEntityNotification
+    unsubscribeEntityNotification,
+    openGoalDetail,
+    shareGoalToMastermind
   }
 )(GoalCard);
